move param clamping to layer base class
